+++ AudioKit/Common/Nodes/Effects/Distortion/AKDecimator.swift
@@ -23,7 +23,7 @@ public struct AKDecimator: AKNode {
     public var avAudioNode: AVAudioNode
     
     /** Decimation (Percent) ranges from 0 to 100 (Default: 50) */
-    public var decimation: Float = 50 {
+    public var decimation: Double = 50 {
         didSet {
             if decimation < 0 {
                 decimation = 0
@@ -31,12 +31,12 @@ public struct AKDecimator: AKNode {
             if decimation > 100 {
                 decimation = 100
             }
-            AudioUnitSetParameter(internalAudioUnit, kDistortionParam_Decimation, kAudioUnitScope_Global, 0, decimation, 0)
+            AudioUnitSetParameter(internalAudioUnit, kDistortionParam_Decimation, kAudioUnitScope_Global, 0, Float(decimation), 0)
         }
     }
     
     /** Rounding (Percent) ranges from 0 to 100 (Default: 0) */
-    public var rounding: Float = 0 {
+    public var rounding: Double = 0 {
         didSet {
             if rounding < 0 {
                 rounding = 0
@@ -44,13 +44,13 @@ public struct AKDecimator: AKNode {
             if rounding > 100 {
                 rounding = 100
             }
-            AudioUnitSetParameter(internalAudioUnit, kDistortionParam_Rounding, kAudioUnitScope_Global, 0, rounding, 0)
+            AudioUnitSetParameter(internalAudioUnit, kDistortionParam_Rounding, kAudioUnitScope_Global, 0, Float(rounding), 0)
         }
     }
     
         
     /** Mix (Percent) ranges from 0 to 100 (Default: 50) */
-    public var mix: Float = 50 {
+    public var mix: Double = 50 {
         didSet {
             if mix < 0 {
                 mix = 0
@@ -58,16 +58,16 @@ public struct AKDecimator: AKNode {
             if mix > 100 {
                 mix = 100
             }
-            AudioUnitSetParameter(internalAudioUnit, kDistortionParam_FinalMix, kAudioUnitScope_Global, 0, mix, 0)
+            AudioUnitSetParameter(internalAudioUnit, kDistortionParam_FinalMix, kAudioUnitScope_Global, 0, Float(mix), 0)
         }
     }
     
     /** Initialize the effect node */
     public init(
         _ input: AKNode,
-        decimation: Float = 50,
-        rounding: Float = 0,
-        mix: Float = 50) {
+        decimation: Double = 50,
+        rounding: Double = 0,
+        mix: Double = 50) {
             
             self.decimation = decimation
             self.rounding = rounding
@@ -82,8 +82,8 @@ public struct AKDecimator: AKNode {
             // Since this is the Decimator, mix it to 100% and use the final mix as the mix parameter
             AudioUnitSetParameter(internalAudioUnit, kDistortionParam_DecimationMix, kAudioUnitScope_Global, 0, 100, 0)
             
-            AudioUnitSetParameter(internalAudioUnit, kDistortionParam_Decimation, kAudioUnitScope_Global, 0, decimation, 0)
-            AudioUnitSetParameter(internalAudioUnit, kDistortionParam_Rounding, kAudioUnitScope_Global, 0, rounding, 0)
-            AudioUnitSetParameter(internalAudioUnit, kDistortionParam_FinalMix, kAudioUnitScope_Global, 0, mix, 0)
+            AudioUnitSetParameter(internalAudioUnit, kDistortionParam_Decimation, kAudioUnitScope_Global, 0, Float(decimation), 0)
+            AudioUnitSetParameter(internalAudioUnit, kDistortionParam_Rounding, kAudioUnitScope_Global, 0, Float(rounding), 0)
+            AudioUnitSetParameter(internalAudioUnit, kDistortionParam_FinalMix, kAudioUnitScope_Global, 0, Float(mix), 0)
     }
 }
